utils/requirements.txt
publicsuffixlist
dnspython
